requirements.txt
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
python-decouple==3.8
python-telegram-bot==20.4
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
